Text-Voice/requirements.txt
requests
pyttsx3
